refactor: restructure Feed module
Feed now has two subpackages: api & impl. Code is currently in impl.
