+++ app/src/main/kotlin/com/potatosheep/kite/app/nav/TopLevelScreen.kt
@@ -57,7 +57,7 @@ import com.potatosheep.kite.core.designsystem.LocalBackgroundColor
 import com.potatosheep.kite.core.translation.R.string as Translation
 import com.potatosheep.kite.feature.about.impl.nav.navigateToAbout
 import com.potatosheep.kite.feature.bookmark.impl.nav.navigateToBookmark
-import com.potatosheep.kite.feature.feed.nav.FeedScreen
+import com.potatosheep.kite.feature.feed.impl.nav.FeedScreen
 import com.potatosheep.kite.feature.image.nav.navigateToImage
 import com.potatosheep.kite.feature.library.nav.HomeScreen
 import com.potatosheep.kite.feature.post.nav.navigateToPost
