+++ feature/feed/impl/src/main/AndroidManifest.xml
