+++ settings.gradle.kts
@@ -28,8 +28,8 @@ include(":core:network")
 include(":core:model")
 include(":core:data")
 include(":feature")
-include(":feature")
-include(":feature:feed")
+include(":feature:feed:impl")
+include(":feature:feed:api")
 include(":core:ui")
 include(":core:designsystem")
 include(":feature:subscription")
